requirements.txt
pandas==2.2.2
numpy==1.26.4
google-cloud-bigquery==3.25.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
faker==30.0.0
matplotlib==3.9.2
seaborn==0.13.2
python-dotenv==1.0.1
